requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-environ>=0.11,<1.0
pillow>=10.0,<12.0
whitenoise>=6.6,<7.0
gunicorn>=21.2,<23.0
psycopg2-binary>=2.9,<3.0
flutterwave
